Revert renaming of Remote to Device
